order-processing-system/ecs-processor/requirements.txt
boto3==1.34.14
psycopg2-binary==2.9.9
